arachnida/scorpion/requirements.txt
requests
numpy
pandas
matplotlib
beautifulsoup4
exifread
pillow
